CORE-4355: Move JarReader and JarVerifier into packaging-verify
In trying to use `JarReader` for reading format version 2, I found that it is tied to verification. It performs signature verification that cannot be turned off.

I attempted to split the functionality but it resulted in a lot of duplicate code within packaging-verify (read jar, then verify at every usage).

As a result I'm moving it back into packaging-verify and making it internal. It is a good fit for packaging-verify which is what it was built for.
