+++ libs/packaging/packaging-verify/src/main/kotlin/net/corda/libs/packaging/verify/internal/cpb/CpbV1Verifier.kt
@@ -1,7 +1,7 @@
 package net.corda.libs.packaging.verify.internal.cpb
 
 import net.corda.libs.packaging.CpkDependencyResolver
-import net.corda.libs.packaging.JarReader
+import net.corda.libs.packaging.verify.JarReader
 import net.corda.libs.packaging.PackagingConstants.CPB_FORMAT_ATTRIBUTE
 import net.corda.libs.packaging.PackagingConstants.CPB_NAME_ATTRIBUTE
 import net.corda.libs.packaging.PackagingConstants.CPB_VERSION_ATTRIBUTE
@@ -19,7 +19,7 @@ import java.util.jar.Manifest
 /**
  * Verifies CPB format 1.0
  */
-class CpbV1Verifier internal constructor (private val packageType: String, jarReader: JarReader): CpbVerifier {
+internal class CpbV1Verifier(private val packageType: String, jarReader: JarReader): CpbVerifier {
     private val name = jarReader.jarName
     private val manifest: Manifest = jarReader.manifest
     private val cpkVerifiers: List<CpkV1Verifier>
