+++ libs/packaging/packaging-verify/src/main/kotlin/net/corda/libs/packaging/verify/internal/cpi/CpiV1Verifier.kt
@@ -1,12 +1,12 @@
 package net.corda.libs.packaging.verify.internal.cpi
 
-import net.corda.libs.packaging.JarReader
+import net.corda.libs.packaging.verify.JarReader
 import net.corda.libs.packaging.verify.internal.cpb.CpbV1Verifier
 
 /**
  * Verifies CPI format 1.0
  */
-class CpiV1Verifier(jarReader: JarReader): CpiVerifier {
+internal class CpiV1Verifier(jarReader: JarReader): CpiVerifier {
     // CPB and CPI format 1 are same
     private val verifier = CpbV1Verifier("CPI", jarReader)
